requirements.txt
djangorestframework-simplejwt
django
djangorestframework
drf-yasg